Integrate Cloudflare/Chromium Zlib
Integrate Cloudflare's optimized Zlib version from:
 https://github.com/cloudflare/zlib
Integrate Chromium's optimized Zlib version from:
 https://chromium.googlesource.com/chromium/src/third_party/zlib

Cloudflare's and Chromium's Zlib have to be configured at build
time with '--with-additional-zlib={cloudflare,chromium}'.

At runtime, it is possible to choose any of the additionally configured
implementations with the new command line option:
  -XX:ZlibImplementation=<cloudflare|chromium|bundled|system>

'bundled' chooses the original, bundled version of Zlib or the system
wide version if there is no bundled one (i.e the JDK was configured
with '--with-zlib=system').
'system' chooses the Zlib version from your system library path.

It is also possible to switch the implementation of Inflate/Deflate
selectively with:
  -XX:ZlibImplementationDeflate=<cloudflare|chromium|bundled|system>
  -XX:ZlibImplementationInflate=<cloudflare|chromium|bundled|system>

The default for all three option is 'bundled'.